Add therapist APIs and currency helper and it's implements.

+++ database/migrations/2020_04_02_095500_currencies_table.php
@@ -18,7 +18,7 @@ class CurrenciesTable extends Migration
         Schema::create('currencies', function (Blueprint $table) {
             $table->id();
             $table->string('code')->unique();
-            $table->float('exchange_rate');
+            $table->decimal('exchange_rate', 8, 4);
             $table->bigInteger('country_id')->unsigned();
             $table->foreign('country_id')->references('id')->on('countries')->onDelete('cascade');
             $table->timestamps();
